# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E3Di & Jaan LED Screen IoT Ad-Network Platform API",
    "description": "Comprehensive REST & IoT API documentation for the E3Di Smart Pole, Digital Billboard Network, Edge Gateways, CCTV surveillance, and Emergency SOS infrastructure.",
    "version": "2.0.0",
    "contact": {
      "name": "E3Di Platform Support",
      "url": "https://e3di.org"
    }
  },
  "servers": [
    {
      "url": "https://api.e3di.org",
      "description": "Production LED CPU API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server (Port 5000)"
    },
    {
      "url": "http://localhost:5001",
      "description": "Alternative Local Server (Port 5001)"
    }
  ],
...
  "paths": {
    "/api/device/{deviceId}": {
      "get": {
        "tags": ["IoT Device & Screens"],
        "summary": "Fetch Screen Playlist & Device State",
        "description": "Called by Android TV Boxes, Raspberry Pis, or Web players to fetch approved active ad playlists, time windows, audio configuration, and device status.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "ethree-pole1" }
        ],
        "responses": {
          "200": {
            "description": "Active playlist and screen configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/PlaylistItem" }
                    },
                    "current": { "$ref": "#/components/schemas/PlaylistItem" },
                    "next": { "$ref": "#/components/schemas/PlaylistItem" },
                    "screen": { "$ref": "#/components/schemas/DeviceScreenState" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/device/{deviceId}/frame": {
      "post": {
        "tags": ["IoT Device & Screens"],
        "summary": "Upload Device Screen Live Frame",
        "description": "IoT screen player uploads a base64 or binary snapshot of what is currently displaying for remote monitoring and proof-of-play.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["frame"],
                "properties": {
                  "frame": { "type": "string", "description": "Data URL or base64 encoded image frame" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Frame received successfully" },
          "400": { "description": "No frame data provided" }
        }
      },
      "get": {
        "tags": ["IoT Device & Screens"],
        "summary": "Get Latest Screen Frame",
        "description": "Fetches the most recent live frame captured by the IoT screen device.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Latest frame data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "frame": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "No frame available or stream offline" }
        }
      }
    },
    "/api/device/{deviceId}/audio": {
      "patch": {
        "tags": ["IoT Device & Screens"],
        "summary": "Remote Audio Control",
        "description": "Adjust volume level (0-100) or toggle sound output on the physical LED screen.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "soundEnabled": { "type": "boolean" },
                  "volume": { "type": "integer", "minimum": 0, "maximum": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Audio configuration updated" },
          "404": { "description": "Screen not found" }
        }
      }
    },
    "/api/device/{deviceId}/k100c": {
      "post": {
        "tags": ["IoT Device & Screens"],
        "summary": "Dispatch BLE Command to K100C Hardware Controller",
        "description": "Sends brightness, power switch, or display mode command via ESP32 BLE cloud bridge to the NovaStar/Colorlight/K100C LED sending card.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "power": { "type": "boolean", "example": true },
                  "brightness": { "type": "integer", "minimum": 0, "maximum": 100, "example": 85 },
                  "mode": { "type": "string", "example": "normal" },
                  "command": { "type": "string", "example": "SET_BRIGHTNESS:85" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Command queued and dispatched via WebSocket/BLE" }
        }
      }
    },
    "/api/device/{deviceId}/k100c/poll": {
      "get": {
        "tags": ["IoT Device & Screens"],
        "summary": "ESP32 WiFi Poll for K100C Commands",
        "description": "Onboard ESP32 microcontroller polls this endpoint periodically over WiFi to retrieve latest hardware state and BLE commands.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Latest hardware state and commands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceId": { "type": "string" },
                    "power": { "type": "boolean" },
                    "brightness": { "type": "integer" },
                    "mode": { "type": "string" },
                    "command": { "type": "string" },
                    "serverTime": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/device/{deviceId}/k100c/heartbeat": {
      "post": {
        "tags": ["IoT Device & Screens"],
        "summary": "ESP32 BLE Heartbeat & Telemetry",
        "description": "ESP32 IoT module reports BLE connection health, signal status, and uptime to the server.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bleConnected": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Heartbeat registered" }
        }
      }
    },
    "/api/device/lookup/{domain}": {
      "get": {
        "tags": ["IoT Device & Screens"],
        "summary": "Lookup Device ID by Domain",
        "parameters": [
          { "name": "domain", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Device ID found" },
          "404": { "description": "No screen mapped to this domain" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "PlaylistItem": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "scheduleId": { "type": "string" },
          "title": { "type": "string" },
          "url": { "type": "string" },
          "durationSeconds": { "type": "integer" },
          "hasWatermark": { "type": "boolean" },
          "date": { "type": "string" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" }
        }
      },
      "DeviceScreenState": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "string" },
          "name": { "type": "string" },
          "location": { "type": "string" },
          "status": { "type": "string" },
          "soundEnabled": { "type": "boolean" },
          "volume": { "type": "integer" }
        }
      },
...
    }
  }
}